refactor(ai-help): extract Markdown component
